requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
pandas==2.1.1
numpy==1.24.3
scipy==1.11.3
statsmodels==0.14.0
requests==2.31.0
python-dotenv==1.0.0
datasets
